Added proper mobile view. Information view now changes to a different layout when mobile is used. Adjusted box width and padding for form-view and information-view. Split trip display components into separate files.
